refactor(transform): isolate shaker module rewrites
